requirements.txt
openai-whisper==20231117
sounddevice==0.4.7
numpy==1.26.4
scipy==1.13.1
pynput==1.7.7
pyperclip==1.9.0
pyautogui==0.9.54
